+++ ruma-events/src/presence.rs
@@ -3,10 +3,12 @@
 //! The only content valid for this event is `PresenceEventContent.
 
 use js_int::UInt;
-pub use ruma_common::presence::PresenceState;
 use ruma_events_macros::{Event, EventContent};
 use ruma_identifiers::UserId;
 use serde::{Deserialize, Serialize};
+
+#[deprecated = "use `presence::PresenceState` from `ruma` or `ruma-common` instead"]
+pub use ruma_common::presence::PresenceState;
 
 /// Presence event.
 #[derive(Clone, Debug, Event)]